Prevent deleting a staff user

+++ routes/admin.js
@@ -278,6 +278,17 @@ router.post('/manage/users/edit/:userId', (req, res) => {
                 })
             }
             if (requestedAction === 'DELETE_USER') {
+                // Check whether demotion is the currentUser
+                if (req.currentUser.id === user.id) {
+                    res.cookie('notifs', `ERR_DELETE然シテ${user.name}然シテCannot delete yourself.`, NotificationCookieOptions)
+                    return res.redirect('/admin/manage/users/edit/' + user.id)
+                }
+
+                if (user.is_admin === true) {
+                    res.cookie('notifs', `ERR_DELETE然シテ${user.name}然シテCannot delete privileged user. Please demote the account before attempting a delete.`, NotificationCookieOptions)
+                    return res.redirect('/admin/manage/users/edit/' + user.id)
+                }
+
                 return User.destroy({
                     where: {
                         'id': targetUserId,
